"def" parameter of config_file::get is now optional
The "def" parameter of config_file::get is now optional. It uses C++11 uniform initialization syntax to automatically fill in the default value of type deduced in the template arguments.

Because of this development, config_file::get<T> is now less verbose to call than the config_file::get_* functions, which have been deprecated.